feat: Backend additions, Portal implementations, Branch Search Feature, Portal split into Menu

+++ MemberSystem/src/main/java/com/isacariotsystems/MemberSystem/entity/User.java
@@ -72,7 +72,7 @@ public class User implements UserDetails {
 
     private String address;
 
-    private float socialScore; 
+    private double socialScore; 
 
     private String firstName;
